+++ src/main/java/io/supertokens/pluginInterface/Storage.java
@@ -17,11 +17,7 @@
 
 package io.supertokens.pluginInterface;
 
-import com.google.gson.JsonObject;
 import io.supertokens.pluginInterface.exceptions.StorageQueryException;
-import io.supertokens.pluginInterface.sqlStorage.SQLStorage;
-
-import javax.annotation.Nullable;
 
 public interface Storage {
 
@@ -37,12 +33,6 @@ public interface Storage {
     // load tables and create connection pools
     void initStorage();
 
-    // no need for transaction like feature here.
-    // get stored appId. If nothing is stored, return null
-    String getAppId() throws StorageQueryException;
-
-    // no need for transaction like feature here.
-    void setAppId(String appId) throws StorageQueryException;
 
     // used by the core to do transactions the right way.
     STORAGE_TYPE getType();
@@ -52,29 +42,11 @@ public interface Storage {
 
     void close();
 
-    void createNewSession(String sessionHandle, String userId, String refreshTokenHash2,
-                          JsonObject userDataInDatabase, long expiry, JsonObject userDataInJWT, long createdAtTime)
-            throws StorageQueryException;
-
-    // return number of rows else throw UnsupportedOperationException
-    int getNumberOfSessions() throws StorageQueryException;
-
-    int deleteSession(String[] sessionHandles) throws StorageQueryException;
-
-    String[] getAllSessionHandlesForUser(String userId) throws StorageQueryException;
-
-    void deleteAllExpiredSessions() throws StorageQueryException;
-
     KeyValueInfo getKeyValue(String key) throws StorageQueryException;
 
     void setKeyValue(String key, KeyValueInfo info) throws StorageQueryException;
 
     void setStorageLayerEnabled(boolean enabled);
-
-    SQLStorage.SessionInfo getSession(String sessionHandle) throws StorageQueryException;
-
-    int updateSession(String sessionHandle, @Nullable JsonObject sessionData, @Nullable JsonObject jwtPayload)
-            throws StorageQueryException;
 
     boolean canBeUsed(String configFilePath);
 }
